requirements.txt
torch
torchvision
opencv-python
tqdm
ultralytics
Pillow
kagglehub
scikit-learn
